requirements.txt
discord.py[voice]>=2.3.0
gTTS>=2.4.0
edge-tts>=6.1.0
PyNaCl>=1.5.0
python-dotenv>=1.0.0
